[update]: Texts updated and download jobs button added (#97)
