+++ clicap/src/components/Sidebar/sidebar.css
