--- clicap/src/components/Sidebar/sidebar.scss
@@ -1,92 +1,0 @@
-/* .sidebar {
-  flex: 1;
-  border-right: 0.5px solid rgb(230, 227, 227);
-  min-height: 100vh;
-  background-color: white;
-  max-width: 15vw;
-
-  .top {
-    height: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    .logo {
-      font-size: 25px;
-      font-weight: bold;
-      color: #3985cb;
-    }
-  }
-
-  hr {
-    height: 0;
-    border: 0.5px solid rgb(230, 227, 227);
-  }
-
-  .center {
-    padding-left: 10px;
-
-    ul {
-      list-style: none;
-      margin: 0;
-      padding: 0;
-
-      .title {
-        font-size: 13px;
-        font-weight: bold;
-        color: #999;
-        margin-top: 15px;
-        margin-bottom: 5px;
-      }
-
-      li {
-        display: flex;
-        align-items: center;
-        padding: 5px;
-        cursor: pointer;
-
-        &:hover {
-          background-color: #94c8f8;
-        }
-
-        .icon {
-          font-size: 20px;
-          color: #3985cb;
-        }
-
-        span {
-          font-size: 20px;
-          font-weight: 600;
-          color: #888;
-          margin-left: 10px;
-        }
-      }
-    }
-  }
-
-  .bottom {
-    display: flex;
-    align-items: center;
-    margin: 10px;
-
-    .colorOption {
-      width: 20px;
-      height: 20px;
-      border-radius: 5px;
-      border: 1px solid #3985cb;
-      cursor: pointer;
-      margin: 5px;
-
-      &:nth-child(1) {
-        background-color: whitesmoke;
-      }
-      &:nth-child(2) {
-        background-color: #333;
-      }
-      &:nth-child(3) {
-        background-color: darkblue;
-      }
-    }
-  }
-}
- */
